backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
paho-mqtt>=1.6.1
google-generativeai>=0.5.0
openai>=1.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
